face-debugger/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
redis==5.1.1
anthropic==0.37.1
httpx==0.27.2
python-dotenv==1.0.1
